Split Cosmos WASM indexer and Cosmos RPC integration
