+++ rust/main/chains/hyperlane-cosmos/src/utils.rs
@@ -12,7 +12,7 @@ use tracing::warn;
 use hyperlane_core::{ChainCommunicationError, ChainResult, Indexed, LogMeta};
 
 use crate::grpc::{WasmGrpcProvider, WasmProvider};
-use crate::rpc::{CosmosWasmIndexer, ParsedEvent, WasmIndexer};
+use crate::rpc::{CosmosWasmRpcProvider, ParsedEvent, WasmRpcProvider};
 
 type FutureChainResults<T> = Vec<JoinHandle<(ChainResult<Vec<(T, LogMeta)>>, u32)>>;
 
@@ -43,15 +43,17 @@ pub(crate) async fn get_block_height_for_lag(
 
 pub(crate) fn parse_logs_in_range<T: PartialEq + Send + Sync + Debug + 'static>(
     range: RangeInclusive<u32>,
-    indexer: Box<CosmosWasmIndexer>,
+    provider: Box<CosmosWasmRpcProvider>,
     parser: for<'a> fn(&'a Vec<EventAttribute>) -> ChainResult<ParsedEvent<T>>,
     label: &'static str,
 ) -> FutureChainResults<T> {
     range
         .map(|block_number| {
-            let indexer = indexer.clone();
+            let provider = provider.clone();
             tokio::spawn(async move {
-                let logs = indexer.get_logs_in_block(block_number, parser, label).await;
+                let logs = provider
+                    .get_logs_in_block(block_number, parser, label)
+                    .await;
                 (logs, block_number)
             })
         })
